requirements.txt
boto3==1.23.5
fsspec[s3]
numpy==1.20.2
pandas==1.1.5
pyarrow==3.0.0
PyYAML==5.3.1
pytorch-lightning==1.2.5
torch-tb-profiler==0.1.0
scikit-learn==0.24.1
scipy==1.5.0
torch==1.8.1
umap-learn==0.5.1
scanpy==1.7.2
s3fs
jedi
jupyter
jupyterlab
pylint
pytest
pytest-rerunfailures
protobuf==3.19.4